requirements.txt
python-dotenv==0.14.0
python-telegram-bot==12.8
